+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+* Fix `shape.intersects()` behavior when a node is dragged
+
 ## 7.2.1
 
 * Fix correct rendering of `Konva.Label` when heigh of text is changed